migrations and unused columns refactoring

+++ database/migrations/tenant/2022_02_18_003727_add_base_currency_to_products_table.php
@@ -14,8 +14,13 @@ class AddBaseCurrencyToProductsTable extends Migration
     public function up()
     {
         Schema::table('products', function (Blueprint $table) {
-            $table->dropColumn('num_of_sale');
-            $table->string('base_currency', 10)->after('unit_price');
+            if (Schema::hasColumn('products', 'num_of_sale')) {
+                $table->dropColumn('num_of_sale');
+            }
+
+            if (!Schema::hasColumn('products', 'base_currency')) {
+                $table->string('base_currency', 10)->after('unit_price');
+            }
         });
     }
